+++ tests/UnitTests/Marvin.Migrations.UnitTests/ScriptMigrationsProviderTests.cs
@@ -10,7 +10,7 @@ namespace Marvin.Migrations.UnitTests
     public class ScriptMigrationsProviderTests
     {
         [Fact]
-        public void GetMigrations_FromDirectory_Ok()
+        public void GetMigrations_FromDirectory_WithoutPrefix_Ok()
         {
             var dbProvider = Mock.Of<IDbProvider>();
 
@@ -50,7 +50,27 @@ namespace Marvin.Migrations.UnitTests
         }
         
         [Fact]
-        public void GetMigrations_FromAssembly_Ok()
+        public void GetMigrations_FromDirectory_WithPrefix_Ok()
+        {
+            var dbProvider = Mock.Of<IDbProvider>();
+
+            var migrationsProvider = new ScriptMigrationsProvider();
+            var path = Path.Combine(Directory.GetCurrentDirectory(), "Scripts");
+            migrationsProvider.FromDirectory(path, "prefix");
+
+            var migrations = migrationsProvider.GetMigrations(dbProvider).ToList();
+            
+            Assert.Equal(1, migrations.Count);
+            
+            Assert.True(migrations[0] is ScriptMigration);
+            Assert.Equal(new DbVersion(0,1), migrations[0].Version);
+            Assert.True(String.IsNullOrEmpty(migrations[0].Comment));
+            Assert.Equal("prefix", ((ScriptMigration)migrations[0]).UpScript);
+            Assert.True(String.IsNullOrEmpty(((ScriptMigration)migrations[0]).DownScript));
+        }
+        
+        [Fact]
+        public void GetMigrations_FromAssembly_WithoutPrefix_Ok()
         {
             var dbProvider = Mock.Of<IDbProvider>();
 
@@ -59,7 +79,7 @@ namespace Marvin.Migrations.UnitTests
 
             var migrations = migrationsProvider.GetMigrations(dbProvider).ToList();
             
-            Assert.Equal(4, migrations.Count);
+            Assert.Equal(5, migrations.Count);
             
             Assert.True(migrations[0] is ScriptMigration);
             Assert.Equal(new DbVersion(1,0), migrations[0].Version);
@@ -86,6 +106,32 @@ namespace Marvin.Migrations.UnitTests
             Assert.True(String.IsNullOrEmpty(migrations[3].Comment));
             Assert.Equal("up", ((ScriptMigration)migrations[3]).UpScript);
             Assert.True(String.IsNullOrEmpty(((ScriptMigration)migrations[3]).DownScript));
+            
+            
+            Assert.True(migrations[4] is ScriptMigration);
+            Assert.Equal(new DbVersion(2,0), migrations[4].Version);
+            Assert.True(String.IsNullOrEmpty(migrations[4].Comment));
+            Assert.Equal("prefix", ((ScriptMigration)migrations[4]).UpScript);
+            Assert.True(String.IsNullOrEmpty(((ScriptMigration)migrations[4]).DownScript));
+        }
+        
+        [Fact]
+        public void GetMigrations_FromAssembly_WithPrefix_Ok()
+        {
+            var dbProvider = Mock.Of<IDbProvider>();
+
+            var migrationsProvider = new ScriptMigrationsProvider();
+            migrationsProvider.FromAssembly(Assembly.GetExecutingAssembly(), "PreMigration");
+
+            var migrations = migrationsProvider.GetMigrations(dbProvider).ToList();
+            
+            Assert.Equal(1, migrations.Count);
+            
+            Assert.True(migrations[0] is ScriptMigration);
+            Assert.Equal(new DbVersion(2,0), migrations[0].Version);
+            Assert.True(String.IsNullOrEmpty(migrations[0].Comment));
+            Assert.Equal("prefix", ((ScriptMigration)migrations[0]).UpScript);
+            Assert.True(String.IsNullOrEmpty(((ScriptMigration)migrations[0]).DownScript));
         }
     }
 }